+++ custom_components/crestron/binary_sensor.py
@@ -11,10 +11,10 @@ from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
-from homeassistant.const import STATE_ON, STATE_OFF, CONF_NAME, CONF_DEVICE_CLASS
+from homeassistant.const import STATE_ON, CONF_NAME, CONF_DEVICE_CLASS
 import homeassistant.helpers.config_validation as cv
 
-from .const import HUB, DOMAIN, VERSION, CONF_JOIN, CONF_IS_ON_JOIN, CONF_BINARY_SENSORS
+from .const import HUB, DOMAIN, VERSION, CONF_IS_ON_JOIN, CONF_BINARY_SENSORS
 from .helpers import get_hub
 
 _LOGGER = logging.getLogger(__name__)