+++ custom_components/crestron/sensor.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Callable
+from typing import Any
 import voluptuous as vol
 import logging