+++ custom_components/crestron/switch.py
@@ -15,7 +15,7 @@ from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
-from homeassistant.const import STATE_ON, STATE_OFF, CONF_NAME, CONF_DEVICE_CLASS
+from homeassistant.const import STATE_ON, CONF_NAME, CONF_DEVICE_CLASS
 from .const import (
     HUB,
     DOMAIN,